COM-20 updated table
