+++ apps/app/src/components/tables/tests/filter-toolbar.tsx
@@ -12,108 +12,98 @@ import { useAction } from "next-safe-action/hooks";
 import { toast } from "sonner";
 
 interface FilterToolbarProps {
-	isEmpty?: boolean;
+  isEmpty?: boolean;
 }
 
-export function FilterToolbar({ isEmpty }: FilterToolbarProps) {
-	const router = useRouter();
-	const pathname = usePathname();
-	const searchParams = useSearchParams();
-	const t = useI18n();
-	const [isPending, startTransition] = useTransition();
-	const [inputValue, setInputValue] = useState(
-		searchParams?.get("search") ?? "",
-	);
+export function FilterToolbar({ isEmpty = false }: FilterToolbarProps) {
+  const router = useRouter();
+  const pathname = usePathname();
+  const searchParams = useSearchParams();
+  const t = useI18n();
+  const [isPending, startTransition] = useTransition();
+  const [inputValue, setInputValue] = useState(
+    searchParams?.get("search") ?? "",
+  );
 
-	const refreshTests = useAction(refreshTestsAction, {
-		onSuccess: () => {
-			toast.success(t("tests.actions.refresh_success"));
+  const refreshTests = useAction(refreshTestsAction, {
+    onSuccess: () => {
+      toast.success(t("tests.actions.refresh_success"));
       window.location.reload();
-		},
-		onError: () => {
-			toast.error(t("tests.actions.refresh_error"));
-		},
-	});
+    },
+    onError: () => {
+      toast.error(t("tests.actions.refresh_error"));
+    },
+  });
 
-	const refreshTestsClick = () => {
-		refreshTests.execute();
-	};
+  const refreshTestsClick = () => {
+    refreshTests.execute();
+  };
 
-	const createQueryString = useCallback(
-		(params: Record<string, string | null>) => {
-			const newSearchParams = new URLSearchParams(searchParams?.toString());
+  const createQueryString = useCallback(
+    (params: Record<string, string | null>) => {
+      const newSearchParams = new URLSearchParams(searchParams?.toString());
 
-			for (const [key, value] of Object.entries(params)) {
-				if (value === null) {
-					newSearchParams.delete(key);
-				} else {
-					newSearchParams.set(key, value);
-				}
-			}
+      for (const [key, value] of Object.entries(params)) {
+        if (value === null) {
+          newSearchParams.delete(key);
+        } else {
+          newSearchParams.set(key, value);
+        }
+      }
 
-			return newSearchParams.toString();
-		},
-		[searchParams],
-	);
+      return newSearchParams.toString();
+    },
+    [searchParams],
+  );
 
-	const [debouncedValue] = useDebounce(inputValue, 300);
+  const [debouncedValue] = useDebounce(inputValue, 300);
 
-	useEffect(() => {
-		startTransition(() => {
-			router.push(
-				`${pathname}?${createQueryString({
-					search: debouncedValue || null,
-					page: null,
-				})}`,
-			);
-		});
-	}, [debouncedValue, createQueryString, pathname, router]);
+  useEffect(() => {
+    startTransition(() => {
+      router.push(
+        `${pathname}?${createQueryString({
+          search: debouncedValue || null,
+          page: null,
+        })}`,
+      );
+    });
+  }, [debouncedValue, createQueryString, pathname, router]);
 
-	return (
-		<div className="flex flex-row items-center justify-between gap-2 mb-4">
-			<div className="flex flex-1 items-center gap-2 min-w-0">
-				{!isEmpty && (
-					<div className="relative flex-1 md:max-w-sm">
-						<Search className="absolute left-2 top-2.5 h-4 w-4 text-muted-foreground" />
-						<Input
-							placeholder={t("tests.filters.search")}
-							className="pl-8"
-							value={inputValue}
-							onChange={(e) => setInputValue(e.target.value)}
-						/>
-					</div>
-				)}
+  return (
+    <div className="flex flex-row items-center justify-between gap-2 mb-4">
+      <div className="flex flex-1 items-center gap-2 min-w-0">
+        {!isEmpty && (
+          <div className="relative flex-1 md:max-w-sm">
+            <Search className="absolute left-2 top-2.5 h-4 w-4 text-muted-foreground" />
+            <Input
+              placeholder={t("tests.filters.search")}
+              className="pl-8"
+              value={inputValue}
+              onChange={(e) => setInputValue(e.target.value)}
+            />
+          </div>
+        )}
+      </div>
 
-				<div className="md:hidden">
-					{
-						<Button asChild variant="action" onClick={refreshTestsClick}>
-							{t("tests.actions.refresh")}
-						</Button>
-					}
-				</div>
-			</div>
-
-			<div className="hidden md:flex items-center gap-2">
-				{inputValue && (
-					<Button
-						variant="ghost"
-						size="sm"
-						onClick={() => {
-							setInputValue("");
-							router.push(pathname);
-						}}
-						disabled={isPending}
-					>
-						<X className="h-4 w-4 mr-2" />
-						{t("common.actions.clear")}
-					</Button>
-				)}
-				{
-					<Button asChild variant="action" onClick={refreshTestsClick}>
-						{t("tests.actions.refresh")}
-					</Button>
-				}
-			</div>
-		</div>
-	);
+      <div className="flex items-center gap-2">
+        {inputValue && (
+          <Button
+            variant="ghost"
+            size="sm"
+            onClick={() => {
+              setInputValue("");
+              router.push(pathname);
+            }}
+            disabled={isPending}
+          >
+            <X className="h-4 w-4 mr-2" />
+            {t("common.actions.clear")}
+          </Button>
+        )}
+        <Button variant="action" onClick={refreshTestsClick}>
+          {t("tests.actions.refresh")}
+        </Button>
+      </div>
+    </div>
+  );
 }
